General Cleaning
Updated the window layout so it can’t collapse in on itself like it
used to. Now if it is resized it can’t go smaller than the content
shown on screen.
Also replaced references to eyePhoto with RAMDisk Accelerator in the
menus and removed unused menu items. Made Open actually work just like
the Add files button and the New menu item work like the New RAMDisk
button. Also made the Open panel accept multiple file selections. The
code was set up to do it but there was one line missing to actually
allow it.
Also tweaked the layout just a little to make things line up a little
nicer.
Changed references to addImage to addFile to reflect their actual use.
For my own use I had to update the build settings and decided to update
some deprecated code as well but I left all of it the same for this
commit so it would still build fine on older versions of OS X.